Added SwaggerUiConfiguration to the swagger-ui library
The main reason for doing this is that WebMvcAutoConfiguration is not loaded
the moment any manual mvc configuration is done. For e.g. we use the WebMvcConfigurationAdapter
to configure message converters in Swagger2DocumentationConfiguration and JacksonSwaggerSupport
As a result the spring boot auto configuration will never kick in as it is a conditional bean

Details here :https://github.com/spring-projects/spring-boot/blob/master/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/web/WebMvcAutoConfiguration.java

In any case, we need a way to tell spring to load webjars and resources from class path. Lifted some of the
code from the WebMvcAutoConfiguration to facilitate that.

related to #638 to support spring boot scenarios